// frontend/src/utilities/types/CalculateTotalProductPriceTypes.ts
/* eslint-disable typescript-sort-keys/interface */
/* eslint-disable id-length */

export type PriceVariablesData = {
  backNumberOfColors: number;
  frontNumberOfColors: number;
  productId: number;
  quantityBySizes: {
    XS: number;
    S: number;
    M: number;
    LG: number;
    XL: number;
    '2XL': number;
    '3XL': number;
    '4XL': number;
    '5XL': number;
  };
};